# docs/openapi/openapi.yaml
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Deterministic, highly resilient backend infrastructure for the Lemni Payment-as-a-Service (PaaS) engine.
  title: LEMNI PaaS Core Engine API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /admin/customers:
    post:
      description: Registers a customer under the current merchant, preparing them for subscription charges.
      operationId: BillingController_registerCustomer
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  example: customer@test.com
                  format: email
                  type: string
                metadata:
                  additionalProperties: true
                  example:
                    companyName: LemonInc
                    department: Support
                  type: object
              required:
                - email
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  createdAt:
                    example: '2026-07-04T02:00:00.000Z'
                    format: date-time
                    type: string
                  email:
                    example: customer@test.com
                    format: email
                    type: string
                  id:
                    example: cust_8a9d2ca8b10
                    type: string
                  merchantId:
                    example: merchant-test-123
                    type: string
                  nombaToken:
                    example: null
                    nullable: true
                    type: string
                type: object
          description: Customer successfully registered
        '401':
          description: Unauthorized JWT session token
      security:
        - bearer: []
      summary: Register a customer
      tags:
        - merchant-dashboard/billing
  /admin/plans:
    post:
      description: Registers a new recurring or one-time subscription plan in the system.
      operationId: BillingController_createPlan
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                amount:
                  example: 29.99
                  format: float
                  type: number
                billingModel:
                  default: recurring
                  enum:
                    - recurring
                    - one_time
                    - custom_input
                  example: recurring
                  type: string
                gracePeriodDays:
                  default: 0
                  example: 3
                  format: int32
                  type: integer
                interval:
                  enum:
                    - weekly
                    - monthly
                    - yearly
                  example: monthly
                  type: string
                name:
                  example: Standard Subscription Plan
                  type: string
                trialDays:
                  default: 0
                  example: 7
                  format: int32
                  type: integer
                trialRequireCard:
                  default: false
                  example: true
                  type: boolean
              required:
                - name
                - amount
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  amount:
                    example: 29.99
                    format: float
                    type: number
                  billingModel:
                    example: recurring
                    type: string
                  gracePeriodDays:
                    example: 3
                    type: integer
                  id:
                    example: plan_7a8dca9
                    type: string
                  interval:
                    example: monthly
                    type: string
                  merchantId:
                    example: merchant-test-123
                    type: string
                  name:
                    example: Standard Subscription Plan
                    type: string
                  trialDays:
                    example: 7
                    type: integer
                  trialRequireCard:
                    example: true
                    type: boolean
                type: object
          description: Plan successfully created
        '401':
          description: Unauthorized JWT session token
      security:
        - bearer: []
      summary: Create a billing plan
      tags:
        - merchant-dashboard/billing
  /admin/plans/{id}:
    delete:
      description: Removes a subscription plan if no active or trialing customers are attached to it.
      operationId: BillingController_deletePlan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Plan successfully deleted
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Cannot delete plan due to active or trialing subscriptions
                    type: string
                type: object
          description: Cannot delete plan due to active subscriptions
        '401':
          description: Unauthorized JWT session token
        '404':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Plan not found
                    type: string
                type: object
          description: Plan not found
      security:
        - bearer: []
      summary: Delete a billing plan
      tags:
        - merchant-dashboard/billing
  /admin/subscriptions/{id}/reactivate:
    post:
      description: Resets the billing cycle start date and shifts the subscription back to active status.
      operationId: BillingController_reactivateSubscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  createdAt:
                    example: '2026-07-04T02:00:00.000Z'
                    format: date-time
                    type: string
                  currentPeriodEnd:
                    example: '2026-08-04T02:00:00.000Z'
                    format: date-time
                    type: string
                  customerId:
                    example: cust_8a9d2ca8b10
                    type: string
                  id:
                    example: sub_a28deca9
                    type: string
                  planId:
                    example: plan_7a8dca9
                    type: string
                  status:
                    example: active
                    type: string
                  trialEnd:
                    example: null
                    format: date-time
                    nullable: true
                    type: string
                type: object
          description: Subscription successfully reactivated
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Subscription is not in canceled status or payment method is missing
                    type: string
                type: object
          description: Subscription is not canceled or card is missing
        '401':
          description: Unauthorized JWT session token
        '404':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Subscription or associated plan not found
                    type: string
                type: object
          description: Subscription or associated plan not found
      security:
        - bearer: []
      summary: Reactivate a canceled subscription
      tags:
        - merchant-dashboard/billing
  /admin/transactions:
    get:
      description: Returns a paginated list of all transactions belonging to the authenticated merchant with filtering support.
      operationId: BillingController_getTransactions
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      properties:
                        amount:
                          example: 29.99
                          type: number
                        createdAt:
                          example: '2026-07-04T10:00:00Z'
                          format: date-time
                          type: string
                        customerId:
                          example: cust_abc123
                          type: string
                        id:
                          example: tx_abc123
                          type: string
                        merchantId:
                          example: merchant-test-123
                          type: string
                        nombaRef:
                          example: nomba_ref_123
                          nullable: true
                          type: string
                        status:
                          example: success
                          type: string
                        subscriptionId:
                          example: sub_abc123
                          nullable: true
                          type: string
                      type: object
                    type: array
                  pagination:
                    properties:
                      limit:
                        example: 20
                        type: integer
                      offset:
                        example: 0
                        type: integer
                      total:
                        example: 100
                        type: integer
                    type: object
                type: object
          description: Transaction history retrieved successfully
        '401':
          description: Unauthorized JWT session token
      security:
        - bearer: []
      summary: Retrieve transaction history
      tags:
        - merchant-dashboard/billing
...
servers: []
...